requirements.txt
colorama
clint @ git+https://github.com/helviojunior/clint.git
kaitaistruct>=0.10
hexdump>=3.3
tabulate>=0.9.0
requests
python-magic>=0.4.27
python-magic-bin>=0.4.14 ; platform_system == "Windows"
cryptography==41.0.4
pyOpenSSL==23.2.0
asn1crypto>=1.5.1
psutil>=5.9.6
pyyaml>=5.0